Refactor provider passport projection wiring packages
