+++ backend/src/main/java/com/alligator/market/backend/provider/config/ProviderFeatureWiringConfig.java
@@ -2,7 +2,7 @@ package com.alligator.market.backend.provider.config;
 
 import com.alligator.market.backend.provider.config.adapter.moex.iss.MoexIssProviderConfig;
 import com.alligator.market.backend.provider.config.application.passport.query.list.PassportListServiceWiringConfig;
-import com.alligator.market.backend.provider.config.readmodel.passport.projection.startup.ProviderPassportProjectionStartupRunnerWiringConfig;
+import com.alligator.market.backend.provider.config.application.passport.projection.startup.ProviderPassportProjectionStartupRunnerWiringConfig;
 import com.alligator.market.backend.provider.config.registry.ProviderRegistryWiringConfig;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Import;
