+++ backend/src/main/java/com/alligator/market/backend/provider/config/application/passport/projection/service/ProviderPassportProjectionServiceWiringConfig.java
@@ -1,7 +1,7 @@
-package com.alligator.market.backend.provider.config.readmodel.passport.projection.service;
+package com.alligator.market.backend.provider.config.application.passport.projection.service;
 
 import com.alligator.market.backend.provider.application.passport.projection.port.ProviderPassportProjectionWritePort;
-import com.alligator.market.backend.provider.config.readmodel.passport.projection.jooq.ProviderPassportProjectionWritePortWiringConfig;
+import com.alligator.market.backend.provider.config.application.passport.projection.jooq.ProviderPassportProjectionWritePortWiringConfig;
 import com.alligator.market.backend.provider.config.registry.ProviderRegistryWiringConfig;
 import com.alligator.market.backend.provider.application.passport.projection.ProviderPassportProjectionService;
 import com.alligator.market.domain.provider.registry.ProviderRegistry;
